Unbreak 'Archive and Clear Database'
The Archive and Clear Database function suffered from trying to create
new tables in cadarchive while holding locks on other tables, which
caused an error.

Change the script so when creating the archive copy of the database
it assumes that locks aren't needed while copying over the database
schema and only while copying the actual user data.
 * Lock the archive_master table and insert the log line
 * drop all locks
 * Create duplicate tablename_timestamp tables in cadarchive
 * Lock all source and destination tables
 * Copy over the contents
 * Clear out the main database for next event